convert demo into example, translate to TS
