+++ examples/container/index.css
@@ -13,5 +13,4 @@ body {
 
 #paper .joint-paper {
     border: 1px solid #E2E2E2;
-    margin: 0;
 }
