+++ examples/container/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description" content="The JointJS Container and Embedding demo serves as a template to help bring your idea to life in no time.">
+    <title>Container and Embedding | JointJS</title>
+  </head>
+  <body>
+    <div id="paper"></div>
+    <script src="dist/bundle.js"></script>
+  </body>
+</html>
