--- packages/joint-core/demo/container/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8"/>
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <meta name="description" content="The JointJS Container and Embedding demo serves as a template to help bring your idea to life in no time.">
-
-        <title>Container and Embedding | JointJS</title>
-
-        <link rel="stylesheet" type="text/css" href="css/container.css" />
-    </head>
-    <body>
-
-        <div id="paper"></div>
-
-        <script src="../../build/joint.js"></script>
-        <script src="src/joint.shapes.container.js"></script>
-        <script src="src/index.js"></script>
-
-    </body>
-</html>
